requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
drf-yasg>=1.21
PyYAML>=6.0
